issue #12
* Fixed the handling of large integers
* Fixed the handling of uninitialized variables
* Fixed the sprintf() function (and related) to behave like GNU AWK (gawk)

+++ src/main/java/org/jawk/frontend/AwkParser.java
@@ -4043,9 +4043,9 @@ public class AwkParser {
 
 	private final class Integer_AST extends ScalarExpression_AST implements NonStatement_AST {
 
-		private Integer I;
+		private Long I;
 
-		private Integer_AST(Integer I) {
+		private Integer_AST(Long I) {
 			this.I = I;
 		}
 
@@ -5090,7 +5090,7 @@ public class AwkParser {
 		// can report accurate line numbers upon errors
 
 		AST addINTEGER(String integer) {
-			return new Integer_AST(Integer.parseInt(integer));
+			return new Integer_AST(Long.parseLong(integer));
 		}
 
 		AST addDOUBLE(String dbl) {
